Convert sanitize-node
